pull deformation interpolator changes into main
