+++ src/backend/update_places_functor.cpp
@@ -54,51 +54,15 @@ void declare_config(UpdatePlacesFunctor::Config& config) {
   name("UpdatePlacesFunctor::Config");
   field(config.pos_threshold_m, "pos_threshold_m", "m");
   field(config.distance_tolerance_m, "distance_tolerance_m", "m");
-  field(config.num_control_points, "num_control_points");
-  field(config.control_point_tolerance_s, "control_point_tolerance_s", "s");
+  field(config.deformation_interpolator, "deformation_interpolator");
   field(config.merge_proposer, "merge_proposer");
   field(config.layer, "layer");
 }
 
-struct AttributeMap {
-  std::vector<NodeAttributes*> attributes;
-
-  void push_back(NodeAttributes* attrs) { attributes.push_back(attrs); }
-
-  void sort() {
-    std::sort(
-        attributes.begin(), attributes.end(), [](const auto& lhs, const auto& rhs) {
-          return lhs->last_update_time_ns < rhs->last_update_time_ns;
-        });
-  }
-};
-
-size_t pgmoNumVertices(const AttributeMap& map) { return map.attributes.size(); }
-
-kimera_pgmo::traits::Pos pgmoGetVertex(const AttributeMap& map,
-                                       size_t i,
-                                       kimera_pgmo::traits::VertexTraits* traits) {
-  const auto& attrs = map.attributes[i];
-  if (traits) {
-    traits->stamp = attrs->last_update_time_ns;
-  }
-
-  return attrs->position.cast<float>();
-}
-
-uint64_t pgmoGetVertexStamp(const AttributeMap& map, size_t i) {
-  return map.attributes[i]->last_update_time_ns;
-}
-
-void pgmoSetVertex(AttributeMap& map,
-                   size_t i,
-                   const kimera_pgmo::traits::Pos& pos,
-                   const kimera_pgmo::traits::VertexTraits&) {
-  map.attributes[i]->position = pos.cast<double>();
-}
-
 UpdatePlacesFunctor::UpdatePlacesFunctor(const Config& config)
-    : config(config::checkValid(config)), merge_proposer(config.merge_proposer) {}
+    : config(config::checkValid(config)),
+      merge_proposer(config.merge_proposer),
+      deformation_interpolator(config.deformation_interpolator) {}
 
 UpdateFunctor::Hooks UpdatePlacesFunctor::hooks() const {
   auto my_hooks = UpdateFunctor::hooks();
@@ -161,79 +125,6 @@ size_t UpdatePlacesFunctor::updateFromValues(const LayerView& view,
   return num_changed;
 }
 
-size_t UpdatePlacesFunctor::interpFromValues(const LayerView& view,
-                                             SharedDsgInfo& dsg,
-                                             const UpdateInfo::ConstPtr& info) const {
-  if (!info->deformation_graph) {
-    return 0;
-  }
-
-  const auto this_robot = GlobalInfo::instance().getRobotPrefix().id;
-
-  std::map<char, AttributeMap> nodes;
-  for (const auto& node : view) {
-    size_t robot_id = this_robot;
-    if (info->node_to_robot_id) {
-      auto iter = info->node_to_robot_id->find(node.id);
-      if (iter == info->node_to_robot_id->end()) {
-        LOG(WARNING) << "Node " << NodeSymbol(node.id) << " does not belong to robot";
-      } else {
-        robot_id = iter->second;
-      }
-    }
-
-    const auto prefix = kimera_pgmo::GetVertexPrefix(robot_id);
-    auto robot_attrs = nodes.find(prefix);
-    if (robot_attrs == nodes.end()) {
-      robot_attrs = nodes.emplace(prefix, AttributeMap{}).first;
-    }
-
-    auto& attrs = node.attributes();
-    auto cache_iter = cached_pos_.find(node.id);
-    if (cache_iter == cached_pos_.end()) {
-      cache_iter = cached_pos_.emplace(node.id, attrs.position).first;
-    } else if (attrs.is_active) {
-      // update cache if node is still active
-      cache_iter->second = attrs.position;
-    }
-
-    // set the position of the node to the original position before deformation
-    attrs.position = cache_iter->second;
-    robot_attrs->second.push_back(&attrs);
-  }
-
-  auto& dgraph = *info->deformation_graph;
-  for (auto& [prefix, attributes] : nodes) {
-    if (!dgraph.hasVertexKey(prefix)) {
-      continue;
-    }
-
-    const auto& control_points = dgraph.getInitialPositionsVertices(prefix);
-    if (control_points.size() < config.num_control_points) {
-      continue;
-    }
-
-    attributes.sort();  // make sure attributes are sorted by timestamp
-    std::vector<std::set<size_t>> vertex_graph_map_deformed;
-    kimera_pgmo::deformation::deformPoints(attributes,
-                                           vertex_graph_map_deformed,
-                                           attributes,
-                                           prefix,
-                                           control_points,
-                                           dgraph.getVertexStamps(prefix),
-                                           *dgraph.getValues(),
-                                           config.num_control_points,
-                                           config.control_point_tolerance_s,
-                                           nullptr);
-  }
-
-  for (const auto& node : view) {
-    dsg.graph->setNodeAttributes(node.id, node.attributes().clone());
-  }
-
-  return 0;
-}
-
 void UpdatePlacesFunctor::call(const DynamicSceneGraph& unmerged,
                                SharedDsgInfo& dsg,
                                const UpdateInfo::ConstPtr& info) const {
@@ -250,18 +141,7 @@ void UpdatePlacesFunctor::call(const DynamicSceneGraph& unmerged,
 
   size_t num_changed = 0;
   if (!info->places_values || info->places_values->size() == 0) {
-    num_changed = interpFromValues(view, dsg, info);
-
-    std::vector<NodeId> to_remove;
-    for (const auto& [node_id, pos] : cached_pos_) {
-      if (!unmerged.hasNode(node_id)) {
-        to_remove.push_back(node_id);
-      }
-    }
-
-    for (const auto& node_id : to_remove) {
-      cached_pos_.erase(node_id);
-    }
+    deformation_interpolator.interpolateNodePositions(unmerged, *dsg.graph, info, view);
   } else {
     num_changed = updateFromValues(view, dsg, info);
   }
